# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book sale",
        "description": "Простой сервис по пробаже книг",
        "version": "1"
    },
    "paths": {
        "/authors": {
            "get": {
                "tags": [
                    "authors"
                ],
                "summary": "Получение списка authors",
                "description": "Получение списка authors",
                "operationId": "33b4ca0e528a8b6298d9942781a0c85a",
                "parameters": [
                    {
                        "parameter": "page",
                        "name": "page",
                        "in": "query",
                        "description": "Текущая страница",
                        "required": false,
                        "deprecated": false,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "paginator": {
                                            "$ref": "#/components/schemas/Paginator"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1"
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "format": "first name",
                                                        "example": "Oleg"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "format": "last name",
                                                        "example": "Oleg"
                                                    },
                                                    "rank": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": "1.24"
                                                    },
                                                    "avatar_url": {
                                                        "type": "string",
                                                        "format": "url",
                                                        "example": "http://rodriguez.com/aut-similique-laudantium-qui-assumenda-ducimus-quisquam"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "authors"
                ],
                "summary": "Создание автора",
                "description": "Создание автора",
                "operationId": "d998cf74685b2e9609bcd48ab986f678",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "format": "first name",
                                        "example": "Oleg"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "last name",
                                        "example": "Oleg"
                                    },
                                    "rank": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "1.24"
                                    },
                                    "avatar_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "http://rodriguez.com/aut-similique-laudantium-qui-assumenda-ducimus-quisquam"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1"
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "format": "first name",
                                                        "example": "Oleg"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "format": "last name",
                                                        "example": "Oleg"
                                                    },
                                                    "rank": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": "1.24"
                                                    },
                                                    "avatar_url": {
                                                        "type": "string",
                                                        "format": "url",
                                                        "example": "http://rodriguez.com/aut-similique-laudantium-qui-assumenda-ducimus-quisquam"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/authors/{id}": {
            "get": {
                "tags": [
                    "authors"
                ],
                "summary": "Получение одного автора",
                "description": "Получение одного автора",
                "operationId": "5317eb225dd93ee043c24cfacda9f26d",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "query",
                        "description": "Идентификатор автора",
                        "required": true,
                        "deprecated": false,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "format": "int",
                                                    "example": "1"
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "format": "first name",
                                                    "example": "Oleg"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "format": "last name",
                                                    "example": "Oleg"
                                                },
                                                "rank": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": "1.24"
                                                },
                                                "avatar_url": {
                                                    "type": "string",
                                                    "format": "url",
                                                    "example": "http://rodriguez.com/aut-similique-laudantium-qui-assumenda-ducimus-quisquam"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "authors"
                ],
                "summary": "Изменение данных автора",
                "description": "Изменение данных автора",
                "operationId": "54818ef139366ae08ef5747f7c9d9a36",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "query",
                        "description": "Идентификатор автора",
                        "required": true,
                        "deprecated": false,
                        "allowEmptyValue": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "format": "first name",
                                        "example": "Oleg"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "last name",
                                        "example": "Oleg"
                                    },
                                    "rank": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "1.24"
                                    },
                                    "avatar_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "http://rodriguez.com/aut-similique-laudantium-qui-assumenda-ducimus-quisquam"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1"
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "format": "first name",
                                                        "example": "Oleg"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "format": "last name",
                                                        "example": "Oleg"
                                                    },
                                                    "rank": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": "1.24"
                                                    },
                                                    "avatar_url": {
                                                        "type": "string",
                                                        "format": "url",
                                                        "example": "http://rodriguez.com/aut-similique-laudantium-qui-assumenda-ducimus-quisquam"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "authors"
                ],
                "summary": "Удаление автора",
                "description": "Удаление автора",
                "operationId": "c8d794bbc8f06854b38ce720c4395db8",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "query",
                        "description": "Идентификатор автора",
                        "required": true,
                        "deprecated": false,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Paginator": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "format": "int",
                        "example": "1"
                    },
                    "first_page_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=1"
                    },
                    "from": {
                        "type": "integer",
                        "format": "int",
                        "example": "1"
                    },
                    "last_page": {
                        "type": "integer",
                        "format": "int",
                        "example": "380"
                    },
                    "last_page_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=5"
                    },
                    "next_page_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=3"
                    },
                    "path": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=2"
                    },
                    "per_page": {
                        "type": "integer",
                        "format": "int",
                        "example": "30"
                    },
                    "to": {
                        "type": "integer",
                        "format": "int",
                        "example": "30"
                    },
                    "total": {
                        "type": "integer",
                        "format": "int",
                        "example": "11371"
                    }
                },
                "type": "object"
            }
        }
    },
...
}